cleanup: test suites
